musica, cambio de escena
la musica ya no se corta al pasar de escena, cuando se pasa a la escena de menu ya se mueve el persona y ya esta el texto en estadisticas pero aun no se actualiza

+++ Final/Assets/script/mainManager.cs
@@ -9,7 +9,7 @@ public class mainManager : MonoBehaviour
 
     public int financiamiento;
     public int financiamientoValor;
-    public int interes;
+    public double interes;
     public bool seguro;
 
     public bool agricultura;
